# contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Kosto-Vet API
  version: 0.5.0
  description: Demo-release API contract for the FastAPI backend. Every operation explicitly declares its access class. Disabled future capabilities return FEATURE_DISABLED and do not imitate success.
  license:
    name: Proprietary
servers:
  - url: https://api.kosto-vet.ru
    description: Production API
  - url: http://localhost:8000
    description: Local FastAPI
...
paths:
...
  /api/v1/payments/robokassa/result:
    post:
      tags: [Payments]
      summary: Receive and verify Robokassa ResultURL notification
      operationId: handleRobokassaResult
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/RobokassaResultNotification"
      responses:
        "200":
          description: Notification committed; body is `OK{InvId}`.
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid signature, amount, invoice id, or payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    CustomerSession:
      type: apiKey
      in: cookie
      name: kv_customer_access
      description: Short-lived customer access session. HttpOnly, Secure, SameSite=Lax.
    CustomerRefreshSession:
      type: apiKey
      in: cookie
      name: kv_customer_refresh
      description: Rotating customer refresh credential. Stored only as a hash server-side.
    AdminSession:
      type: apiKey
      in: cookie
      name: kv_access
      description: Short-lived staff access session with server-side role and permission checks.
    AdminRefreshSession:
      type: apiKey
      in: cookie
      name: kv_refresh
      description: Rotating staff refresh credential. Stored only as a hash server-side.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Signed session-bound CSRF token required together with cookie auth on unsafe methods.
    OrderAccessToken:
      type: apiKey
      in: header
      name: X-Order-Access-Token
      description: Signed opaque guest-order credential. Returned once, stored only as a hash, and never put in URLs or logs.
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        meta:
          type: object
          additionalProperties: true
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
...
    RobokassaResultNotification:
      type: object
      required: [OutSum, InvId, SignatureValue]
      properties:
        OutSum:
          type: string
          pattern: '^[0-9]+(\.[0-9]{1,2})?$'
        InvId:
          type: integer
          minimum: 1
        SignatureValue:
          type: string
          minLength: 32
          maxLength: 128
        PaymentMethod:
          type: string
        IncSum:
          type: string
        Shp_order:
          type: string
      additionalProperties: true
...